+++ bobbin_frontend/.prettierignore
@@ -13,3 +13,4 @@ next.config.js
 next.config.mjs
 postcss.config.js
 tailwind.config.ts
+coverage/
